fix tooltip and remove old plot
