+++ packages/ui/src/components/ObsPlot/ObsTooltipRow.tsx
@@ -4,18 +4,11 @@ import { ReactElement } from "react";
 type ObsTooltipRowProps = {
   children: ReactElement;
   label?: string;
-  labelWidth?: string;
   valueWidth?: string;
   truncateValue?: boolean;
 };
 
-function ObsTooltipRow({
-  children,
-  label,
-  labelWidth,
-  valueWidth,
-  truncateValue = false,
-}: ObsTooltipRowProps) {
+function ObsTooltipRow({ children, label, valueWidth, truncateValue = false }: ObsTooltipRowProps) {
   const truncateLine = truncateValue
     ? {
         whiteSpace: "nowrap",
@@ -26,7 +19,7 @@ function ObsTooltipRow({
 
   return (
     <tr style={{ verticalAlign: "top" }}>
-      <td width={labelWidth}>
+      <td>
         <Typography
           variant="subtitle2"
           fontSize={13}
